src/mcp/backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
sqlite3  # Built-in Python module

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Email Validation and Sending
email-validator==2.1.0
pydantic[email]==2.5.0

# Google Gemini AI
google-generativeai==0.3.2

# HTTP Client (for external API calls)
httpx==0.25.2
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Date and Time
python-dateutil==2.8.2

# Async Support
asyncio  # Built-in Python module
aiofiles==23.2.1

# Development and Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1

# CORS Support (already included in FastAPI)
# python-cors==4.3.1  # Not needed as FastAPI has built-in CORS

# Logging and Monitoring
structlog==23.2.0

# Type Hints
typing-extensions==4.8.0

# JSON handling
orjson==3.9.10  # Faster JSON for FastAPI

# Validation
pydantic==2.5.0

# ASGI Server
gunicorn==21.2.0  # For production deployment

# File handling
pillow==10.1.0  # For future image processing needs

# Caching (for future use)
redis==5.0.1  # Optional for caching

# Monitoring (optional)
prometheus-client==0.19.0  # For metrics
